requirements.txt
apache-airflow==1.10.2
tenacity

pytest
pytest-cov

flake8

